+++ spec/system/user_show_spec.rb
@@ -19,34 +19,6 @@ RSpec.describe 'User', type: :system do
     @comment4 = Comment.create(author: @tom, post: @post1, text: 'Hey , Jerry')
   end
 
-  describe 'User visits the index page' do
-    scenario 'I can see the username of all other users' do
-      visit users_path
-      expect(page).to have_content 'List of Users'
-      expect(page).to have_content(@tom.name)
-      expect(page).to have_content(@jerry.name)
-    end
-
-    scenario 'I can see the profile picture for each user' do
-      visit users_path
-
-      expect(page).to have_selector("img[src='#{@tom['photo']}']")
-      expect(page).to have_selector("img[src='#{@jerry['photo']}']")
-    end
-
-    scenario 'I can see the number of posts each user has written.' do
-      visit users_path
-
-      expect(page).to have_content "Number of posts: #{@tom.posts_counter}"
-      expect(page).to have_content "Number of posts: #{@jerry.posts_counter}"
-    end
-
-    scenario 'When I click on a user, I am redirected to that user show page.' do
-      visit users_path
-      click_on 'Tom'
-      expect(current_path).to eq(user_path(@tom.id))
-    end
-  end
   describe 'User visits the user#show page' do
     scenario 'I can see the users profile picture.' do
       visit user_path(@tom.id)
